Progress being made
I added more scripts to help with the player health and I tried fixing the camera movement, because once the player would be destroyed the camera would be destroyed as well since it was a child of the player. The camera won't stay behind the player, so not sure if I should keep it that way.

I still need to add the health hud to show the damage being taken and health displayed.